prior sensors working
the previous issue not completely solved. If a train is stationary with
facing 1 and flips heading, the enter signal won't be triggered and
there's a chance the route is
blocked by the time the enter signal is triggered.
Since the train planned to pass it won't be able to flip the heading.

Short term hack could be to allow flip heading when it's started, and
stop the train, but it probably won't work with ble train. Ideally, we
should do the checks that we do on _on_target_entered before flipping
the train heading.