fix route locking issue
with the new layout part of the commit, a situation arises where the train
unlocks track that is part of the next leg, which deletes the trainname from the
lock metadata.

Fix it by promoting the locked metadata to a list of trainnames, so
double-locking is allowed (but should only happen with the same
trainname). After unlocking, one trainname is erased.